uploaded new text engine
Finally updated the text engine! It means that now you can track for floats and strings as well(before you had an option only for integers). There is also an option to create a slider/arrowbox and choosing between the float and integer form. 

Change also in text input engine:  
- you can edit the float texts precision via the macro FLOAT_TEXT_PRECISION. By default it is set to 3(be careful as big precisions may cause problem with setprecision function, as it will show closest numbers)
- the numerical variables are not changed during the input mode(while writing), changes apply after you quit or enter the clipboard