New clipboard type update
Added a new clipboard type calles ADJUSTABLE. It means that you have to define the fixed size of text, it means that you can pass more chars to the textbox than its length but less than selected adjust size. It is supported only for normal clipboard and clipboard with fixed sprite width (it means its not possible for textboxes with dynamic variables - or not recommended).