Fast fix upload
Fixed a bug where IconButtons were not cropped correctly due to a bad function called for setting the Width and Height of its collider sprite. (Instead of setting the real width and height I was setting the local one).
Adjustable texts now end with 3 dots if they are longer than the selected size.